refactor: split function to spec part and logic part 2
